# services/news-feed-engine/api/openapi.yaml
openapi: 3.1.0
info:
  title: ElevatedIQ News Feed Engine API
  description: |
    AI-powered news aggregation platform API for multi-platform content ingestion,
    analysis, and video summary generation. Supports white-label multi-tenant deployment.
  version: 1.0.0
  contact:
    name: ElevatedIQ Platform Team
    email: platform@elevatediq.ai
  license:
    name: Proprietary
servers:
  - url: https://api.elevatediq.ai/news-feed/v1
    description: Production
  - url: https://staging-api.elevatediq.ai/news-feed/v1
    description: Staging
  - url: http://127.0.0.1:8096/api/v1
    description: Local Development
...
paths:
...
  /content:
    get:
      tags: [Content]
      summary: List content
      operationId: listContent
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/category'
        - $ref: '#/components/parameters/platform'
        - name: sort
          in: query
          schema:
            type: string
            enum: [published_at, engagement_score, quality_score]
            default: published_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated list of content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
  /content/{id}:
    get:
      tags: [Content]
      summary: Get content by ID
      operationId: getContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Content item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /content/category/{category}:
    get:
      tags: [Content]
      summary: Get content by category
      operationId: getContentByCategory
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Content filtered by category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
  /content/geo/{classification}:
    get:
      tags: [Content]
      summary: Get content by geographic classification
      operationId: getContentByGeo
      parameters:
        - name: classification
          in: path
          required: true
          schema:
            type: string
            enum: [local, regional, national, global]
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Content filtered by geo classification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
  /content/trending:
    get:
      tags: [Content]
      summary: Get trending content
      operationId: getTrendingContent
      parameters:
        - name: range
          in: query
          description: Time range for trending calculation
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d]
            default: 24h
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Trending content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
  /content/search:
    get:
      tags: [Content]
      summary: Search content (semantic)
      operationId: searchContent
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/category'
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /creators/{id}/content:
    get:
      tags: [Creators]
      summary: Get creator's content
      operationId: getCreatorContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: type
          in: query
          schema:
            type: string
            enum: [article, video, post, podcast, newsletter]
      responses:
        '200':
          description: Creator's content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
...
  /analytics/content:
    get:
      tags: [Analytics]
      summary: Get content analytics
      operationId: getContentAnalytics
      security:
        - bearerAuth: []
      parameters:
        - name: range
          in: query
          schema:
            type: string
            enum: [1d, 7d, 30d, 90d]
            default: 7d
        - $ref: '#/components/parameters/category'
        - $ref: '#/components/parameters/platform'
      responses:
        '200':
          description: Content analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentAnalyticsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/content/ingest:
    post:
      tags: [Admin]
      summary: Trigger content ingestion
      operationId: triggerIngestion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionRequest'
      responses:
        '202':
          description: Ingestion job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    category:
      name: category
      in: query
      schema:
        type: string
    platform:
      name: platform
      in: query
      schema:
        type: string
        enum: [youtube, twitter, reddit, rss, internal]
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_more:
          type: boolean
    Content:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
        creator_id:
          type: string
          format: uuid
        platform:
          type: string
          enum: [youtube, twitter, reddit, rss, internal]
        content_type:
          type: string
          enum: [article, video, post, podcast, newsletter]
        title:
          type: string
        description:
          type: string
        original_url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        summary:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        geo_classification:
          type: string
          enum: [local, regional, national, global]
        sentiment:
          type: number
          format: float
          minimum: -1
          maximum: 1
        quality_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        engagement_score:
          type: number
          format: float
        view_count:
          type: integer
        like_count:
          type: integer
        comment_count:
          type: integer
        share_count:
          type: integer
        published_at:
          type: string
          format: date-time
        ai_analysis:
          type: object
          properties:
            key_points:
              type: array
              items:
                type: string
            entities:
              type: array
              items:
                type: object
            topics:
              type: array
              items:
                type: string
        creator:
          $ref: '#/components/schemas/Creator'
    Creator:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        platform:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        tier:
          type: string
          enum: [platinum, gold, silver, bronze, unverified]
        verified_at:
          type: string
          format: date-time
        follower_count:
          type: integer
        engagement_rate:
          type: number
          format: float
...
    IngestionRequest:
      type: object
      properties:
        sources:
          type: array
          items:
            type: string
        platform:
          type: string
        force:
          type: boolean
          default: false
...
    ContentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Content'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ContentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Content'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string

    # AI Schemas
...
    ContentAnalyticsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            time_range:
              type: string
            total_content:
              type: integer
            content_by_category:
              type: array
              items:
                type: object
                properties:
                  category:
                    type: string
                  count:
                    type: integer
                  percentage:
                    type: number
                    format: float
            content_by_platform:
              type: array
              items:
                type: object
                properties:
                  platform:
                    type: string
                  count:
                    type: integer
            content_over_time:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  count:
                    type: integer
            avg_quality_score:
              type: number
              format: float
            avg_engagement_score:
              type: number
              format: float
...